+++ malachite-native/src/integer/conversion/assign_u64.rs
@@ -1,7 +1,7 @@
 use integer::Integer;
 use traits::Assign;
 
-/// Assigns a `u32` to to an `Integer`.
+/// Assigns a `u64` to to an `Integer`.
 ///
 /// Time: worst case O(1)
 ///
@@ -13,11 +13,11 @@ use traits::Assign;
 /// use malachite_native::traits::Assign;
 ///
 /// let mut x = Integer::from(-123);
-/// x.assign(456);
+/// x.assign(456u64);
 /// assert_eq!(x.to_string(), "456");
 /// ```
-impl Assign<u32> for Integer {
-    fn assign(&mut self, other: u32) {
+impl Assign<u64> for Integer {
+    fn assign(&mut self, other: u64) {
         self.sign = true;
         self.abs.assign(other);
     }
